Converted task edit page into reusable component
